ws: add potential use after free on downstream TX errors
In case `send_iov()` fails with a fatal error (such as broken pipe), we
free the client context using `ws_terminate()` but continue using it in
a subsequent call to `uwsd_state_transition()`, eventually leading to a
use-after-free access on the next IO event.

Restructure the function logic to either free the context and return on
error or to schedule a delayed send on retryable `send_iov()` errors.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>