state: invoke IO read callbacks right away if buffered data available
Add facilities to query the buffered data state of connection handles and
use those to decide whether epoll registration is required on state change.

In case buffered data is found to be present and read readiness of a
descriptor required, invoke the associated IO callback right away instead
of relying on epoll notification. This is required for SSL connections
which might have pending application data stored in their context while
the underlying descriptor is fully read and not going to yield another
event.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>