listen: ensure that accepted client sockets are in nonblocking mode
Due to the way pending socket buffer data is handled by state.c, the
accepted socket might never actually get added to the uloop which
implicitly takes care of enabling nonblocking mode on the descriptor,
leading to blocking client socket descriptors being passed to I/O
callbacks which expect nonblocking behaviour.

Mitigate this issue by using accept4() (or a drop-in replacement on
systems where it isn't supported) to produce descriptors which are
nonblocking from the get-go.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>