script: only issue kill() if a script worker was actually started
When a script context is freed before it has been started, e.g. due to a
configuration error, we must ensure to not invoke kill() on the still zero
pid in order to prevent killing the main process.

Solve the issue by checking whether a callback has been assigned, which
only happens after the worker process has been forked.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>