http: rework argument skipping in uwsd_http_reply_buffer_varg()
Since vsnprintf() is specified to leave the passed va_list in undefined
state after it completes, we cannot rely on it skipping the arguments
for us. The current implementation works on glibc but does not behave
as intended on musl libc.

In order to solve this issue, implement the argument skipping manually
by parsing the given format string literal.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>