script: add uwsd.uuid() function
Implement a new uuid() function in the uwsd namespace which can be used
by script handlers to generate UUID values. The uuid function takes an
optional seed value to prime the prng, passing the same seed will yield
the same UUID value.

If no seed is given, the current time is used as seed value.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>