http: use `splice(2)` to transfer backend data when possible
Extend the HTTP frontend to utilize the `splice(2)` syscall in conjunction
with an intermediate buffer pipe in order to transfer backend reply data
downstream to the client socket. The splice based fast path is only used
if both up- and downstream connections are non-SSL and if the upstream
response has a known, predetermined length.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>